Move button component outside of App.tsx
Going to start pulling different logic into different React Routes. We
will need the button component in different routes.
